Extract radar error/toast subsystem into RadarErrorToast
First step of the WeatherRadar refactor: move tile-error tracking and the
"connection instability / rate limit" toast out of the 1200-line class into a
focused RadarErrorToast module. WeatherRadar composes it and injects a layer
redraw via an onRetry callback, so the toast has no knowledge of layers/map.
Behaviour is unchanged; affected white-box tests now reference radar.errorToast.
A TODO remains to continue splitting the other concerns (frames, playback,
polling, reconcile).
