+++ tests/weather-radar-toast-timer.test.js
@@ -15,44 +15,44 @@ describe('WeatherRadar Toast Timer', () => {
         vi.clearAllMocks();
         globalThis.lastRafCb = null;
         radar = new WeatherRadar({});
-        radar.ui.errorToast = { classList: { contains: vi.fn().mockReturnValue(true), add: vi.fn(), remove: vi.fn() }, style: {} };
+        radar.errorToast.ui.errorToast = { classList: { contains: vi.fn().mockReturnValue(true), add: vi.fn(), remove: vi.fn() }, style: {} };
     });
 
     afterEach(() => { vi.restoreAllMocks(); vi.useRealTimers(); });
 
     it('updateTimer aborts if toast is no longer visible', () => {
-        radar.showErrorToast('Title', 'Message', 5);
-        radar.ui.errorToast.classList.contains.mockReturnValue(false);
+        radar.errorToast.showErrorToast('Title', 'Message', 5);
+        radar.errorToast.ui.errorToast.classList.contains.mockReturnValue(false);
         globalThis.lastRafCb();
-        expect(radar.toastAnimationFrame).toBeNull();
+        expect(radar.errorToast.toastAnimationFrame).toBeNull();
     });
 
     it('updateTimer requests next frame if time remains', () => {
         vi.useFakeTimers();
         vi.setSystemTime(1000000);
-        radar.showErrorToast('Title', 'Message', 5);
+        radar.errorToast.showErrorToast('Title', 'Message', 5);
 
         vi.setSystemTime(1001000);
         globalThis.requestAnimationFrame.mockClear();
         globalThis.lastRafCb();
 
-        expect(radar.ui.errorTimer.textContent).toBe('4s');
+        expect(radar.errorToast.ui.errorTimer.textContent).toBe('4s');
         expect(globalThis.requestAnimationFrame).toHaveBeenCalled();
     });
 
     it('updateTimer finishes and hides toast when time is up', () => {
         vi.useFakeTimers();
         vi.setSystemTime(1000000);
-        radar.showErrorToast('Title', 'Message', 5);
+        radar.errorToast.showErrorToast('Title', 'Message', 5);
 
         vi.setSystemTime(1006000);
-        radar.hideErrorToast = vi.fn();
-        radar.rateLimitResetTime = 0;
+        radar.errorToast.hideErrorToast = vi.fn();
+        radar.errorToast.rateLimitResetTime = 0;
 
         globalThis.lastRafCb();
 
-        expect(radar.ui.errorTimer.textContent).toBe('');
-        expect(radar.toastAnimationFrame).toBeNull();
-        expect(radar.hideErrorToast).toHaveBeenCalled();
+        expect(radar.errorToast.ui.errorTimer.textContent).toBe('');
+        expect(radar.errorToast.toastAnimationFrame).toBeNull();
+        expect(radar.errorToast.hideErrorToast).toHaveBeenCalled();
     });
 });
